Defatalize warnings in module and test program
The experimental 'smartmatch' operator will be deprecated in
soon-to-be-released perl-5.38.0.  That deprecation will throw a warning
which, if warnings have been fatalized, will cause a build-time failure.

This patch merely defatalizes the warning.  The warning will still be
omitted (e.g., during './Build test), but this buys time for avoiding
use of smartmatch altogether.